# docs/openapi-v4.1.yaml
openapi: 3.0.3
info:
  title: StarMate English API v4.1
  version: 4.1.2
  description: |
    第一阶段（内部自用）API 合同。
    范围：公开招生入口 + 学生/老师/家长/创始人闭环 + 管理台（平台内侧）+ 审计与AI工具。
servers:
  - url: https://localhost
    description: Cloudflare Pages Functions
security:
  - bearerAuth: []
paths:
  /api/v1/auth/login:
    post:
      summary: 账号登录
      description: 支持演示角色或正式账号密码登录。
      tags: [auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthLoginPayload'
      responses:
        '200':
          description: 登录成功
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponse' }
...
  /api/v1/public/courses:
    get:
      summary: 获取公开课程列表
      tags: [public]
      security: []
      parameters:
        - in: query
          name: institutionId
          schema: { type: string }
        - in: query
          name: grade
          schema: { type: string }
        - in: query
          name: status
          schema: { type: string }
        - in: query
          name: classType
          schema: { type: string }
        - in: query
          name: limit
          schema: { type: integer }
        - in: query
          name: offset
          schema: { type: integer }
      responses:
        '200':
          description: 课程列表
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponse' }
...
  /api/v1/admin/culture-wall:
    get:
      summary: 获取/下载文化墙素材
      tags: [admin]
      parameters:
        - in: query
          name: kind
          schema: { type: string }
        - in: query
          name: placement
          schema: { type: string }
        - in: query
          name: mediaKey
          schema: { type: string }
      responses:
        '200':
          description: 素材列表或流数据
          content:
            application/json:
              schema: { type: object, additionalProperties: true }
    post:
      summary: 新增文化墙素材
      tags: [admin]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                kind:
                  type: string
                  description: photo/video/teacher/feedback
                placement:
                  type: string
                  description: 用于图片位分区，例如 admissions 或 culture-wall
                title:
                  type: string
                description:
                  type: string
                uploader:
                  type: string
                duration:
                  type: string
                sortOrder:
                  type: number
                tags:
                  type: string
                  description: 逗号分隔或 JSON 数组字符串
                file:
                  type: string
                  format: binary
          application/json:
            schema:
              type: object
              properties:
                kind:
                  type: string
                  description: teacher/feedback
                placement:
                  type: string
                title:
                  type: string
                description:
                  type: string
                uploader:
                  type: string
                sortOrder:
                  type: number
                tags:
                  type: array
                  items:
                    type: string
                text:
                  type: string
      responses:
        '200':
          description: 新增结果
          content:
            application/json:
              schema: { type: object }
    delete:
      summary: 删除文化墙素材
      tags: [admin]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CultureWallDeletePayload'
      responses:
        '200':
          description: 删除结果
          content:
            application/json:
              schema: { type: object }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ApiResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: integer
          description: 业务码（0 表示成功）
          example: 0
        message:
          type: string
          example: ok
        data:
          type: object
          additionalProperties: true
        request_id:
          type: string
        trace_id:
          type: string
    AuthLoginPayload:
      type: object
      required: [role]
      properties:
        role: { type: string, enum: [platform, founder, teacher, parent, student] }
        username: { type: string }
        password: { type: string }
...
    CultureWallDeletePayload:
      type: object
      required: [institutionId, id]
      properties:
        institutionId: { type: string }
        id: { type: string }